Final/requirements.txt
flask
celery
redis
#tkinter
#Python 3.11 
#tensorFlow
matplotlib
numpy
requests
scipy
ramdom 
#zipfile
pillow
pytz
cv2